% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.bcaboot}
\alias{autoplot.bcaboot}
\title{Plot BCa bootstrap confidence intervals}
\usage{
\method{autoplot}{bcaboot}(object, ...)
}
\arguments{
\item{object}{A \code{bcaboot} object from \code{\link[=bca_nonpar]{bca_nonpar()}} or \code{\link[=bca_par]{bca_par()}}.}

\item{...}{Ignored.}
}
\value{
A \code{ggplot} object.
}
\description{
Creates a ggplot showing BCa and standard confidence intervals
across coverage levels.
}
\examples{
\dontrun{
data(diabetes, package = "bcaboot")
Xy <- cbind(diabetes$x, diabetes$y)
rfun <- function(Xy) {
  y <- Xy[, 11]; X <- Xy[, 1:10]
  summary(lm(y ~ X))$adj.r.squared
}
set.seed(1234)
result <- bca_nonpar(Xy, 1000, rfun, n_groups = 34, verbose = FALSE)
autoplot(result)
}

}
