% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcaboot.R
\docType{package}
\name{bcaboot}
\alias{bcaboot}
\alias{bcaboot-package}
\title{Automatic Construction of Bootstrap Confidence Intervals}
\description{
Bootstrap confidence intervals depend on three elements: (a) the
cumulative distribution of the bootstrap replications, (b) the
bias-correction, and (c) the acceleration number that measures the
rate of change in the standard deviation of the estimate as the
data changes.  The first two of these depend only on the bootstrap
distribution, and not how it is generated: parametrically or
non-parametrically. Therefore, the only difference in a parametric
bca analysis would lie in the nonparametric estimation of the
acceleration, often a negligible error.
}
