% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bcajack2}
\alias{bcajack2}
\title{Nonparametric BCa via regression (deprecated)}
\usage{
bcajack2(
  x,
  B,
  func,
  ...,
  m = NULL,
  mr,
  pct = 0.333,
  K = 2,
  J = 12,
  alpha = c(0.025, 0.05, 0.1, 0.16),
  verbose = TRUE
)
}
\arguments{
\item{x}{An \eqn{n \times p} data matrix. Rows are observed
\eqn{p}-vectors, assumed independently sampled. If \eqn{p = 1}
then \code{x} can be a vector.}

\item{B}{Number of bootstrap replications (integer).}

\item{func}{Function computing the parameter of interest.
\code{func(x, ...)} should return a scalar for any data matrix.}

\item{...}{Additional arguments passed to \code{func}.}

\item{m}{Number of groups (now \code{n_groups}).}

\item{mr}{Unused (included for compatibility).}

\item{pct}{KL fraction (now \code{kl_fraction}).}

\item{K}{Jackknife repetitions (now \code{n_jack}).}

\item{J}{Jackknife groups (now \code{jack_groups}).}

\item{alpha}{Alpha levels (now use \code{conf.level}).}

\item{verbose}{Logical; show progress bar during bootstrap.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{bcajack2()} was deprecated in bcaboot 1.0 in favour of \code{\link[=bca_nonpar]{bca_nonpar()}}.
}
\keyword{internal}
