% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{bcaplot}
\alias{bcaplot}
\title{BCa plot (deprecated)}
\usage{
bcaplot(vl, ...)
}
\arguments{
\item{vl}{Output of a bcaboot computation function.}

\item{...}{Additional arguments (ignored).}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{bcaplot()} was deprecated in bcaboot 1.0 in favour of
\code{autoplot()}.
}
\keyword{internal}
