% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcaboot.R
\docType{data}
\name{diabetes}
\alias{diabetes}
\title{Blood and other measurements in diabetics}
\format{
This data frame contains the following columns:
\itemize{
\item \strong{x} a matrix with 10 columns
\item \strong{y} a numeric vector
\item \strong{x2} a matrix with 64 columns
}
}
\source{
\url{https://web.stanford.edu/~hastie/Papers/LARS/LeastAngle_2002.pdf}
}
\description{
The \code{diabetes} data frame has 442 rows and 3 columns. These are the
data used in the Efron et al "Least Angle Regression" paper.
}
\details{
The x matrix has been standardized to have unit L2 norm in each column and
zero mean. The matrix x2 consists of x plus certain interactions.
}
\references{
Efron, Hastie, Johnstone and Tibshirani (2003) "Least Angle
Regression" (with discussion) \emph{Annals of Statistics}
}
\keyword{datasets}
