% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gplot.R
\name{ggbcaplot}
\alias{ggbcaplot}
\title{Plots of bca confidence limits}
\usage{
ggbcaplot(vl, main = "", xlab = "coverage", ylab = "limits",
  alpha = c(0.025, 0.05, 0.1, 0.16), ...)
}
\arguments{
\item{vl}{output of \code{bcajack}, \code{bcajack2}, or
\code{bcapar}}

\item{main}{the main caption (can be empty)}

\item{xlab}{x axis label (supplied if not specified)}

\item{ylab}{y axis labels (supplied if not specified)}

\item{...}{further args for ggplot}

\item{two-sided}{coverages are \eqn{1-2\alpha},
e.g. \code{alpha=c(.025,.05)} plots intervals
\code{[.025,.975]} and \code{[.05,.95]}. Default is
\code{alpha=c(.025,.05,.1,.16)} giving coverages
.95,.90,.80,.68}
}
\description{
\code{ggbcaplot} uses the output of \code{bcajack},
\code{bcajack2}, or \code{bcapar} to plot bca and standard
confidence limits for the parameter of interest.
}
\details{
confidence interval endpoints are plotted vertically versus
two-sided coverages \eqn{1-2\alpha}. Bca limits in black,
Standard limits in green (dashed.). If \code{vl$lims} includes
the column "jacksd" of jackknife internal standard deviations
then these are indicated by vertical red bars centered at the
bca limit points.
}
