% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcanon.R
\name{bcanon}
\alias{bcanon}
\title{Compute nonparametric bca bootstrap confidence intervals}
\usage{
bcanon(
  B,
  x,
  func,
  ...,
  m = nrow(x),
  pct = 0.333,
  K = 2,
  J = 12,
  alpha = c(0.025, 0.05, 0.1, 0.16),
  rou = 3,
  catj = 500,
  sw = 0
)
}
\arguments{
\item{B}{number of bootstrap replications}

\item{x}{data matrix of dimension \eqn{n\times p}, rows assumed mutually independent}

\item{func}{R function, where \eqn{func(x, \ldots)} is the statistic of interest}

\item{m}{number of units; if \eqn{m<n} then original units collected into groups each of size \eqn{n/m}; useful if \eqn{n} is very large}

\item{pct}{proportion of original units used in finding gradient}

\item{K}{controls jackknife standard error calculations (how exactly?)}

\item{J}{controls jackknife standard error calculations (how exactly?)}

\item{alpha}{percentiles of coverage probabilities for bootstrap intervals}

\item{rou}{number of digits rounded to}

\item{catj}{counts number of bootstrap replictions; a value of 0 suppresses count}

\item{sw}{if nonzero, adds the replicates \code{tt} and \code{dd} to output; refer to the returned value below}
}
\value{
a named list containing:
\itemize{
\item \strong{call}: the call
\item \strong{lims} : Bca confidence limits (first column) and the standard
limits (third column). The second column, \code{jacksd}, are the
jackknife estimates of Monte Carlo error; \code{pct}, the third
column are the proportion of the replicates \code{tt} less than each
\code{bcalim} value
\item \strong{stats} : Estimates and their jackknife Monte Carlo errors:
\code{theta} = \eqn{\hat{\theta}}; \code{sd}, the bootstrap standard
deviation for \eqn{\hat{\theta}}; \code{a0} the acceleration
estimate; \code{A} the big-A measure of raw acceleration; \code{sese} of
standard error stablity
\item \strong{equiv} : shows the \code{gbca} coverage estimates
$\eqn{\tilde{\alpha}}$ corresponding to the nomimal bca
coverages \eqn{\alpha}; also their jacknife montecarlo standard
deviations.
}
}
\description{
Compute nonparametric bca bootstrap confidence intervals
}
