% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bcapar.R
\name{bcapar}
\alias{bcapar}
\title{Compute parametric bootstrap confidence intervals}
\usage{
bcapar(t0, tt, bb, alpha = c(0.025, 0.05, 0.1, 0.16, 0.5, 0.84, 0.9,
  0.95, 0.975), J = 10, K = 6, trun = 0.001, pct = 0.333, cd = 0,
  func)
}
\arguments{
\item{t0}{Observed estimate of theta, usually by maximum
likelihood.}

\item{tt}{A vector of parametric bootstrap replications of theta of
length \code{B}, usually large, say \code{B = 2000}}

\item{bb}{A \code{B} by \code{p} matrix of natural sufficient vectors, where
\code{p} is the dimension of the exponential family.}

\item{alpha}{Confidence values for which the confidence limits are
calculated}

\item{J, K}{Parameters controlling the jackknife estimates of Monte
Carlo error: \code{J} jackknife folds, with the jackknife standard
errors averaged over \code{K} random divisions of \code{bb}}

\item{trun}{Truncation parameter used in the calculation of the
acceleration \code{a}.}

\item{pct}{Proportion of "nearby" b vectors used in the calculation
of \code{t.}, the gradient vector of theta.}

\item{cd}{If cd is 1 the bca confidence density is also returned
(See Section 11.6 of "Computer Age Statistical Inference",
B. Efron and T. Hastie.)}

\item{func}{Function \code{thetahat = t(b)}. If this is not missing then
output includes \emph{abc} estimates (See
"More accurate confidence intervals in exponential families",
T. DiCiccio and B. Efron, Biometrika 1992 p231-245.)}
}
\value{
a named list of several items:
\describe{
\item{\code{lims}}{bca confidence limits (first column) and the standard
limits (fourth column). Also the abc limits (fifth column) if
\code{func} is provided. The second column, \code{jacksd}, are the
jackknife estimates of Monte Carlo error; \code{pctiles}, the third
column are the proportion of the replicates \code{tt} less than each
\code{bcalim} value}
\item{\code{stats}}{estimates and their jackknife Monte Carlo errors:
\code{thet} = \eqn{\hat{\theta}}; \code{sd}, the bootstrap standard deviation
for \eqn{\hat{\theta}}; \code{a} the acceleration estimate; \code{az} another
acceleration estimate that depends less on extreme values of \code{tt};
\code{z0} the bias-correction estimate; \code{A} the big-A measure of raw
acceleration; \code{sdd} delta method estimate for standard deviation of
\eqn{\hat{\theta}}; \code{mean} the average of \code{tt}}
\item{abcstats}{The abc estimates of \code{a} and \code{z0}, returned if \code{func} was provided}
\item{ustats}{The bias-corrected estimator \code{2 * t0 - mean(tt)}. \code{ustats}
gives \code{ustat}, an estimate \code{sdu} of its sampling error, and jackknife
estimates of monte carlo error for both \code{ustat} and \code{sdu}. Also given
is \code{B}, the number of bootstrap replications}
}
}
\description{
bcapar computes parametric bootstrap confidence
intervals for a real-valued parameter theta in a p-parameter
exponential family. It is described in Section 4 of the
reference below.
}
\examples{
data(diabetes, package = "lars")
X <- diabetes$x
y <- scale(diabetes$y, center = TRUE, scale = FALSE)
lm.model <- lm(y ~ X - 1)
mu.hat <- lm.model$fitted.values
sigma.hat <- sd(lm.model$residuals)
t0 <- summary(lm.model)$adj.r.squared
y.star <- sapply(mu.hat, rnorm, n = 1000, sd = sigma.hat)
tt <- apply(y.star, 1, function(y) summary(lm(y ~ X - 1))$adj.r.squared)
b.star <- y.star \%*\% X
bcapar(t0 = t0, tt = tt, bb = b.star)
}
\references{
DiCiccio T and Efron B (1996). Bootstrap confidence
intervals. Statistical Science 11, 189-228

T. DiCiccio and B. Efron. More accurate confidence intervals in exponential families.
Biometrika (1992) p231-245.

Efron B (1987). Better bootstrap confidence intervals. JASA 82, 171-200

B. Efron and T. Hastie. Computer Age Statistical Inference. Cambridge University Press, 2016.
}
